Initial commit
Setting up a new project based on [1] to create a GitHub Actions action to automatically run SVGO on a SVGs in Pull Requests.

The project will be written in TypeScript and compiled to JavaScript. Additionally, as suggested in [2], the output of the TypeScript compiler will be run through @zeit/ncc to create a runnable file in the repository (until this action is actually published).

Furthermore this commit contains some initial files:

- .editorconfig: Configuration for EditorConfig [3].
- .gitattributes: Git file to specify attributes of files. Primarily added to avoid EOL problems and hide diffs of computer generated files.
- .gitignore: ignore the dependencies (node_modules) folder.
- action.yml: describes the GitHub Action.
- LICENSE: the MIT license that applies to the project.
- package-lock.json: dependencies lockfile.
- package.json: Node manifest file, contains metadata & dependencies
- README.md: placeholder README file for the project.
- tsconfig.js: TypeScript compiler configuration.

--
1. https://github.com/actions/labeler
2. https://help.github.com/en/actions/automating-your-workflow-with-github-actions/creating-a-javascript-action
3. https://editorconfig.org/