test: put getChangedFiles into separate module and test what it returns
The `getChangedFiles` function from src/main.ts was moved to src/github-api.ts in order to be able to test it in isolation. The tests in this commit are slightly more interesting than before.

The test makes use of fixtures (found in test/fixtures) that were created based on real GitHub API responses. This way the `getChangedFiles` functions gets as-close-to-real-as-possible data back from the GitHub client.

The @actions/github and @actions/core modules are mocked (@actions/core is not really necessary yet, just added to try it out). The mocks are not loaded using Jest's automock but manually from the test/mocks folder, mainly because I just prefer it that way. The mocks are written in JavaScript as oposed to TypeScript because I couldn't figure out how to mock a whole type, and e.g. the github.GitHub type has ~28 fields/methods that would need to be mocked explicitly.