test: add fixtures for getFile function
Add a new fixture file for fixture data corresponding to the return
data of the API endpoint used by the `getFile` function.

For consistency, the existing fixtures file, which contains fixture data
for the API endpoint used by the `getPullRequestFiles` function, has
been renamed.

Lastly, the mocking module that uses fixtures has been updated in
accordance with the updates to the fixture files.