feat: encapsulating class for SVGO instance + initial tests
Add an encapsulating class for SVGO which can be used to create an SVG
optimizer with one configuration and subsequently optimize as many SVGs
as is needed.

The test suite for this new module contains basic tests to verify the
class can be instantiated and the optimize method does something
sensible.