refactor: get all Pull Request comments, and return in reverse order
For the ability to enable the action with newer comments the Action will need to look at the newest comments first. Unfortunately, GitHub's API doesn't allow us to fetch the comments in that order [1] and so now we first fetch all comments and then return the list of comments in the reverse order.

--
1. https://developer.github.com/v3/issues/comments/#list-comments-on-an-issue