fix: optimization table bug when not all SVGs are optimized
Fix a bug in the templating system where the optimization table is
incorrect if not all SVGs in the Pull Requests needed to be optimized.

This bug was caused by an assumption that all SVGs would be optimized.
Namely, the original SVG was obtained by simply getting the data from
the list of original SVGs at the index of the optimized SVG.

The bug has been fixed by instead *finding* the original SVG based on
the path of the optimized SVG. Since the path must be equal between
the original and optimized SVG **and** it will be unique, this must
give the correct SVG back. Hence, the bug is fixed.