feat: read comments value as comment teplate
... if it is a string and it does not equal "true". As a result, setting 
the `comments` option to a string automatically enables comments and 
configures the comment template.

Setting it to any of `true` and `"true"` will result in the Action 
leaving comments using the standard comment template.

Setting it to any of `false` and `"false"` will still disable the 
feature.

Not touching the "comments" option still results in no comments.