ci: build and commit changes for lib/index.js
Add a new GitHub Actions workflow file that, for commits on the `develop` branch,
will build lib/index.js and push it back if it changed. This way the lib/index.js
file will always be up-to-date on the `develop` branch no matter what happens.

If the last Pull Request that get's merged doesn't change the lib/index.js file
the Action will not push a commit and exit without an error.