refactor: entrance function for Action
Change the entrance function for the Action from just starting to
process a Pull Request to first detecting the event(Name) of the context
in which the Action is being run.

Currently, the Action detects "push" and "pull_request" events. In the
latter case the Action will run as it always has. In the former the
Action currently fails. This should get an implementation where, similar
to the "pull_request" context, any SVGs in the pushed commit(s) are
optimized.