refactor: initialize action in main
Move the Action initialization (getting configuration etc.) from the
main function in prs.ts to the (new) main function in main.ts. This way
it will be easy to share these variables between a PR and a push run.

This change required some changes to the tests for prs.ts, as well as
skipping some tests that are no longer relevant there (they're skipped
instead of removed here because they will be useful to test the new
main.ts)