chore: update pre-commit hook
Update the pre-commit hook to no longer include changes
to the files in the lib/ directory in a commit (by
default). Instead, we can relay on the GitHub Actions
Workflow that updates lib/ with every commit on the
develop branch.

Also, update the .gitignore file to exclude ALL files
in the lib/ directory except for the exact files we want
to be there (instead of just all .js files).