fix: failing test due to (ts-)jest update
Fix a failing test due to a feature introduced in TS 3.9 nightly that
disallows `delete`ing a property that is not optional. The mocked
actions/github context object did not use optional fields (whereas the
real one does). Hence, the type of the context is now made explicit in
the mock.

Unfortunately, this required one minor change in an existing test suite.
This test suite changes the `number` field of the now-optional
pull_request field, so the test now checks if this optional field is
set.