chore: update git hooks (#244)
* chore: update pre-commit hook

Update the pre-commit hook to no longer include changes
to the files in the lib/ directory in a commit (by default).
Instead, we can relay on the GitHub Actions Workflow that
updates lib/ with every commit on the develop branch.

* chore: update .gitignore

Also, update the .gitignore file to exclude ALL files
in the lib/ directory except for the exact files we want
to be there (instead of just all .js files).

* chore: add pre-push hook

Add pre-push hook through Husky to run tests. Before the
tests are ran all changes will be staged.

* chore: action commit without verify

Update the version of stefanzweifel/git-auto-commit-action that is used
and use the `commit_options` option to ensure that the --no-verify
option is used when committing. This will prevent the pre-commit hook
from stashing the changes in lib/ before committing.