feat: preview of comments/commits with validator
Add support for a preview of what comments and commit
messages will look like for a given configuration with
the config validator tool.

In order to achieve this the ActionConfig object and
templating.ts are used. Because this tool is eventually
intended for the web, I had to refactor those two
modules such that they don't have any external deps.
This is most notable w.r.t. the ActionConfig. The
inputs.ts module has been significantly refactor to no
longer need @actions/core. Instead, the ActionConfig
now expects an object matching an interface from which
inputs can be get using core's `.getInput` function.
(and the other functions in the inputs.ts module have
been moved to src/main.ts as that is the only place
they're used anyway).