feat: add possibility for warnings in messaging
Update the templating logic (src/templating.ts) to be able to include a
list of warnings in messages (excluding commit titles, where they are
ignored).  The templating variable for this is "{{warnings}}".

If there are no warnings this templating variable will result in an
empty string, if there are any warnings a string of the following format
will be inserted:

  WARNINGS:
  - warning 1
  - warning 2
  - etc.

Currently, the Action will only record warnings when trying to get SVGs
from a GitHub repository.

This change required an update to the `CommitData` and `ContextData`
types, both of which now include a field called "warnings" of the type
`string[]` which can be used to simply add a warning represented by a
simple string.