feat: proof of concept on: schedule behaviour
Add a proof of concept of the on: schedule event behaviour of the
Action. It will find all files in the repository, filter out the SVGs
(and the ignored SVGs), optimize them and commit them back to the
default branch.

One unfortunate thing about this is that the @actions/github context in
the on: schedule event does not have the appropriate values to construct
the commit URL using the existing method.

One thing to consider is whether to add an Action option to configure
which branch is being committed to in the on: schedule context.