refactor: getContents in github-api.ts
Refactor out the call to `client.repos.getContent` in github-api.ts
such that it can be obtained in both `getContent` and `getFile`. To
achieve this custom types were introduced.

Additionally, the tests, mocks, and fixtures had to be updated to
accomodate the new usage of `client.repos.getContent`. (which, as a
bonus, provides 100% code coverage.)