fix: add  field to validation example commit data
Add the "warnings" field missing from the EXAMPLE_COMMIT_DATA in the
config validation script.

This field was introduced recently to inform the user about any
warnings explicitly. However, even though the CommitData type was
updated accordingly the example commit data in the script was not.
Likely due to the fact that the script is not typically ran (and also
not inculded in any CI pipeline).