refactor: require ref in github-api.ts
Update the github-api.ts module to require an explicit ref when getting
a file (or "contents"). Update all usage of this module and related
tests accordingly.

The reason for this is that the pull_request context wants to use a
different ref from schedule and push. More specifically, the
pull_request_target event **requires** a custom ref, and the
pull_request also works with that custom ref.