chore: version bump to 2.0.0-alpha
Bump the version on this branch to 2.0.0-alpha as it already has a bunch
of breaking changes. The branch can be used with the `at-next` suffix
through GitHub actions the same way the `v1` suffix works and the
documentation has been updated accordingly. A 2.0.0-alpha tag will not
be added until this branch is stable.