refactor: source & test rewrite
Rewrite the Action source to a more modularized style. The rewrite is
in a Golang-inspired style. The rewrite is not complete yet, but it's
mostly done and in a state where it (should) equals the old behaviour.

The test have been rewritten from scratch entirely. There's now a split
between unit tests (touching one file and one file only, with some
exceptions) and integration tests which test entire folders. Relatedly,
the previously called "integration" test in the CI is now called an
End-to-end (e2e) test.

Some other noteworthy changes:
- Tests now also have an 80 character line limit.
- The `lint:ws` now omits line lenght linting in favor of others.