docs: add repo-token to the events documentation
Add the `repo-token` input the "Options" sections of the events
documentation.

Also, add a column to the "Options" section of the events documentation,
indicating that `repo-token` is a required input.

Note that as of writing this `repo-token` is required for schedule,
workflow_dispatch, and repository_dispatch events even though it isn't
used. This is a bug that should be fixed.