fix: use default input values embedded in source
Refactor package inputs to don't return an error for inputs that are not
required (hence preventing a warning from the Action) AND update it, as
well as the Action manifest, to use the defaults specified in the source
code (rather than the manifest). For clarity the (now mismatching)
"required" value in the manifest has been removed as well.