chore: unpin Node version on `.nvmrc`
Unpin the NodeJS version from v16.6.0 exactly to allowing any v16
instance of NodeJS. In practice this should result in the latest v16
release available on the system being used, which should be sufficient.

This comes after a blog post [1] released today covering a vulnerability
in npm where it is recommended to upgrade to the latest version of
NodeJS, 16.8.0 as of today. Since v16.6.0 is before that, either the
.nvmrc value had to updated or relaxed.

--
1. https://github.blog/2021-09-08-github-security-update-vulnerabilities-tar-npmcli-arborist/#what-you-can-do-to-remediate