docs: fix invalid thollander/actions-comment-pull-request version
The version `v1.0.1` of thollander/actions-comment-pull-request does not
exist (the tag is `1.0.1`, without the `v`). The example now uses an
the latest version of the action as well.

Unfortunately this action does not provide a `v1`-like tag to be used
here. The closest thing to that is recommending the usage of `main`
(following the action's docs) but I'm not comfortable doing that as
`main` _should_ not be used like that. I filed a ticket for this
upstream.