chore: improve existing Contributing Guidelines content
(an attempt to) Simplify and clarify the existing content of the
Contributing Guidelines. General goal of the changes is to use simpler
words, more accurate words, shorter sentences, simpler sentence
structure, and be more explicit.

This commit focusses on the (code) contributor part, i.e. the part from
the "Workflow" section onwards.