chore: add `isStrictMode` option to package inputs
Add `isStrictMode` as a value to the Config interface of the inputs
package and set it to the value of the `strict` input value at runtime
through the newly added `getIsStrictMode` function. The input name and
default value are encoded as constants in the constants package.

The package inputs mocks have been updated accordingly to be able to run
the tests.