test: does the Action change the content of SVG files
Test this by getting the SVG contents in a step before the action is run
and comparing it to the content of the SVG after the action has run.

In the step preceding the Action where the SVGs are read, some character
manipulation is performed to ensures newlines are properly dealt with
when outputting/exporting the raw SVG string.