fix: don't filter push or PR files when running e2e tests
This is, arguably, a temporary fix that allows for running the
end-to-end tests meaningfully - i.e. while actually optimizing
something. Ideally, to test the code we don't need to introduce
additional code. However,  the event variable in question is not
changable within the context of a GitHub Actions run, so another
solution is required.

The current in-code approach is not ideal, but simple. It's not ideal
because the fix is quite "deep" inside the code. It would be better if
this was handled at a higher level (e.g. `src/index.ts`) as it would
make the code easier to understand.