General refactoring (#455)
* refactor: simplify package inputs' index.js

* chore: normalize mocks

Normalize mock names and the use of `__common__.ts`.

* refactor: normalize type definitions

- Alphabetically order types and type attributes.
- Decompose interfaces.
- Reuse types better
- Use `Iterable` where possible.

* refactor: introduce barrel constants.ts files

Add a local `constants.ts` file to each package (that uses constants)
so that each package can import constants from a local file. Currently,
these `constants.ts` files are all barrel files that just import the
constants from package constants and export it again. This makes it
easier to change constants as there's always just one file per package
to change.

* chore: fix, improve, and normalize (unit) test descriptions

* refactor: further improve consistency of types

* refactor: update & improve usage of `constants.ts` files

Expanding the usage of package-local `constants.ts` files to not just be
a barrel file for project-wide constants but also for defining and
sharing local constants.

* chore: remove unused type documentation