refactor: read svgo-version input as string
Update the inputs package to read the svgo-version input as a string and
update all related code to deal with the value that way. Note that the
input is currently restricted to three strings, namely:
- `"1"`
- `"2"` (default)
- `"project"`

Notable is the removal of a test in `getters.test.ts`, this is because
the test used to cover the scenario where a non-numeric string was
provided as input for `svgo-version`. This no longer applies and is now
instead captured by the fact that `svgo.New` outputs an error for
unknown strings. This is covered by the parameterized test case added
for the svgo package's `index.test.ts` unit test suite.