feat: parse SVGO config file based on extension
Add support to parse an SVGO configuration file based on the extension
of the file. I.e., if it's extension is a common YAML extension then it
is parsed as a YAML file, otherwise it is parsed as a JavaScript file.

Note that the existing functionality (parse based on the svgo-version
value) is kept here to ensure perfect backwards compatability, this may
not be necessary though.