feat: import project-local SVGO for svgo-version "project"
Import - but not yet use - the SVGO version of the project using this
Action, as well as distingiush between SVGO v1 and v2 for that import.
Currently, the existing built-in `SVGOptimizer` instance for the package
imported SVGO version is returned. This needs to be refactored so that
the package imported SVGO version is used.