chore: remove impossible if-branch
Remove the if-branch when getting svgo-version input in case the input
is invalid. This is impossible because string inputs are always "valid".
I.e., `core.getInput` always works if provided when running an Action
because GitHub Actions errors if a `with:` value is configured to either
an array or object.