Lint importing and exporting of types in source code (#479)
* chore: lint importing and exporting of types in source code

Update the ESLint configuration to enable linting of importing and
exporting types to only allow `import type` and `export type` when
importing and exporting (resp.) types.

The export-version of this rule requires type information and therefore
the `parserOptions.project` value to be set. For simplicity, these
linting rules where enabled only for the source code, which is what the
existing `tsconfig.json` file is for. If this option is applied at the
top-level it will cause issues with tests and scripts.

* refactor: normalize type imports and exports