docs: update 'Using Any SVGO Version' example
Omit the suggestion to run `npm install svgo` in CI and instead promote
users to add SVGO as a (dev)dependency to their project and only use
`npm ci` to install SVGO as a project-level dependency.

The reasoning behind this is to (ever so slightly) promote security.
When installing dependencies using `npm ci` the checksum of the
downloaded version of SVGO _should_ be checked against the one in the
lockfile (provided it exists), this provides a minimum guarentee on the
integrity of SVGO.