refactor: remove unneeded error check in CommitsClient.listFiles
Remove an unneeded check for an error in the package clients internal
CommitClient class' `listFiles` method. The check isn't needed because
the method returning the error ensures (through the type system) the
returned non-error value is well-formed.

This programming error was detected as a (single) mutant where line 66
of (the old) clients.ts file could be replaced by (only):

    if (true) {

This suggests that the conditional is unnecessary. Inspection of the
code (described above) confirmed this.