ci: don't upload mutation report as Actions artifact
Remove the uploading of the Stryker mutation report as an Actions
artifact as the report is now uploaded to the Stryker dashboard, which
is sufficient for most instances. If something where to happen to the
Stryker dashboard, it is still possible to get the report by running the
mutation tests locally.