test: using file handles when reading files
Update the test suite for the base file-system module's readFile
function to reflect the requirements for basing it on file handles. This
requirements are based on [1] and [2].

The `fs` module mock is updated to include a mocked version of the
`openSync` function.

The `openSync` function shall be used as all existing functions used
from the `fs` module are the `-Sync` versions.

--
1. https://nodejs.org/api/fs.html#fsopensyncpath-flags-mode
2. https://nodejs.org/api/fs.html#fsopenpath-flags-mode-callback