refactor: use file handles when reading files
Replace the check exists-then-read implementation by a open handle-then-
read implementation. The former leaves a small time gap in which, for
example, the file can be deleted. The latter ensures this isn't possible
by holding a claim on the file, while still checking if the file exists
(`openSync` throws an error if the file doesn't exist).