chore: update stashing strategy for pushing
When pushing, in contrast to when committing, all changes should be
stashed in order to create a repository state that most closely
resembles what's being pushed. Hence, indexed files should not be kept
(`--keep-index`) but included in the stash. Additionally, when popping
the stash, the index (i.e. stashed or not) of files should be restored
(`--index`).