chore: setup project build using rollup.js
Set up rollup.js [1] for compiling, bundling, and minifying the source
code into a single standalone JavaScript file.

First and foremost, the at-rollup/plugin-typescript plugin is used to be
able to compile the TypeScript into JavaScript. Following it's docs, the
`tsconfig.json` file is updated to change the compile target.
Additionally, at-rollup/plugin-node-resolve, at-rollup/plugin-commonjs, and
at-rollup/plugin-json are used to make the compilation work (providing
everything necessary to bundle in the (relevant) contents of the
node_modules directory). Lastly, rollup-plugin-terser is used to minify
the build output using Terser [2].

--
1. https://rollupjs.org
2. https://terser.org/