chore(deps): add dependency `eval`
Add a new dependency for evaluating JavaScript configuration files after
reading them as a file, called `eval` [1]. This will replace `node-eval`
[2], which is currently used for that purpose.

Why change? Primarily because `node-eval` was last updated 4 years ago
and `eval` 7 days ago, indicating that it is still actively maintained.
Additionally, it's defaults are more secure (I started looking into this
because I was wondering if anything was leaking into the context of the
JavaScript evaluation).

Note: the last activity in the `node-eval` repo from the maintainer is
also ~4 years ago.

--
1. https://www.npmjs.com/package/eval
2. https://www.npmjs.com/package/node-eval