ci: add step-security/harden-runner to workflows
Add the step-security/harden-runner [1] Action to this project's GitHub
Actions Workflows. The latest version at the time of committing is
pinned.

The purpose of this Action is to first detect "what" _should_ happen
during a typical job run, and then enforce that with specialized config
after that. E.g., detect what HTTP traffic should happen, and then block
all other HTTP traffic in subsequent runs based on the first run.

--
1. https://github.com/step-security/harden-runner