Add step-security/harden-runner to harden GH Actions (#540)
Add the step-security/harden-runner [1] Action to this project's GitHub
Actions Workflows. The latest version at the time of committing is
pinned.

The purpose of this Action is to first detect "what" _should_ happen
during a typical job run, and then enforce that with specialized config
after that. E.g., detect what HTTP traffic should happen, and then block
all other HTTP traffic in subsequent runs based on the first run.

The updating.yml workflow does not yet block any traffic. This is
because it couldn't be triggered in the Pull Request context. It should
be configured later based on runs with `egress-policy: audit`.

--
1. https://github.com/step-security/harden-runner