ci: create GitHub Actions workflow for releases
Create a workflow that can be triggered manually to initiate a release
and is triggered on `main` to release by creating a tag. The latter will
only create tags when there is not yet a tag for the version found in
the manifest (`package.json`) in the commit (see line 86 of
`release.yml`). Additionally, this takes care of keeping the `v3` tag
up-to-date by syncing it with the latest release tag (see line 92 to 94
of `release.yml`).