chore: update the Release Guidelines
Most importantly, stop creating GitHub releases for v2 to avoid v2
releases showing up as latest as well as showing up as the latest
release to use on the GitHub Marketplace (for Actions).

Additionally, this rewrites the release guidelines to be more
straightforward to follow along with. On top of that, this includes some
improvements to the existing descriptions making the release process
both easier to carry out and harder to do wrong. E.g. there's now a
suggested command for updating the version umber in the package
manifest.

Lastly, this fixes a couple of leftover mistakes in the release
guidelines that suggested these guidelines are for the current major
version, whereas in reality they're for v2 of the Action.

Lastly, update the EditorConfig configuration file to not care about
indentation in MarkDown files. This is needed, per the comment, because
sometimes uneven indentation is required for the MarkDown to be "valid".
In particular, in the case of a numbered list, an indentation of 3
spaces is needed to stay in the scope of the current list item if there
should be a blank line in the list item.