Add scheduled workflow to audit dependencies
Add a new GitHub Actions workflow for auditing dependencies. It is
scheduled weekly for an initial trail, and audits dependencies on both
`main`, `main-v1`, and `main-v2`. The latter two are especially of
interest, as a failed job would alert about problems on v1 that are not
present on v3 (since `npm audit` exits with a non-zero exit code if a
problem is detected). The former is more-or-less already covered by
Dependabot/RenovateBot.