ci: don't fail audit jobs fast
Don't fail other audit jobs when one fails. We want to know all audit
result, regardless of some audits failing - whether due to an unexpected
or due to a non-zero audit exit due to a detected error. This is because
an audit error for one branch does not imply one for other branches.
Hence, jobs should fail independently.