ci: use custom audit command for deps check
Use custom audit command when auditing main branches. This supports
using a custom auditing tool per branch. In particular, for now the
intend is to try using better-npm-audit [1] on main-v1 and main-v2 to be
able to ignore known vulnerabilities that cannot be resolved.

If nothing else, `npm run audit` can always be mapped back to
`npm audit`.

--
1. https://www.npmjs.com/package/better-npm-audit