chore: update Node.js version for dev to ^18 and npm to ^8.1.2
Update the Node.js version to be used for development to 18 (or higher)
by updating `.nvmrc` and `package.json`. Update the npm version to be
used during development to 8.1.2 (or higher) by updating `package.json`.
The Contributing Guidelines are updated accordingly. (Also updated the
formatting of "Node.js" and "npm" in the Contributing guidelines in line
with the official formatting, based on the website `<title>`s.)

Regarding the npm version, this version was chosen specifically because
it add support for string values for the `--lockfile-version` flag which
is used to specify the lockfile version used in this project. In .nvmrc
this is specified as a string. From the changes notes for v8.1.2:

> allow `--lockfile-version` config to be string and coerce to number