ci: update installed Node.js version for build & tests
Change the Node.js version that is installed during the code validation
workflow from a version matching `./action.yml` to simply the version
present in `./.nvmrc`. The thinking behind the pre-existing approach was
that the version should match `./action.yml` to more closely match a
"real" Action execution. However, during a real execution of the Action
the Node.js version being installed might also be different from the
version specified in the Action manifest, so I don't think it actually
makes sense to require the installed version to match the Action runtime
version.

This is changed in particular here because the Node.js and npm version
used for development are being upgraded to a version not matching what's
present in `./action.yml`, which leads to the CI failure on the `npm ci`
step because of the `"engine"` requirements specified in `package.json`
(in combination with the `engine-strict=true` set in `./.npmrc`).