chore: add local e2e testing instruction
Add instructions to the Contributing Guidelines for how to run the e2e
tests locally using nektos/act [1] - a tool to "Run your GitHub Actions
locally". Using Docker, nektos/act will execute GitHub Actions workflows
locally. So, using nektos/act, the e2e tests for this project, which
consistute executing the Action and checking its outputs, can be run
locally.

Additionally, a npm script/command is added to run the e2e tests without
having to know the name of the job in the GitHub Actions Workflow. Full
option names are used for readability (since it doesn't have to be
retyped). The command assumes `act` is installed on the system and
available on the PATH.

--
1. https://github.com/nektos/act