chore(deps): pin CodeQL action to latest known working version
This pins the CodeQL action to the latest version that worked on this
repository. The current latest version (v1.1.14, release 3 days ago as
of this commit) produces an error:

  Error: Cannot read property 'ml_powered_queries_enabled' of undefined

Given the state of this branch (v1 is almost end of life (EOL)), and the
fact that a security fix should be merged as soon as possible, we'll
start by just pinning the Action to a known working version. If anyone
is interested in upgrading the Action to a more recent version, a
contribution for this is welcome. Note that upgrading beyond v1 is not
_necessary_ even though it is also EOL - the EOL of the CodeQL Action is
after that of (v1 of) this Action.