ci: don't fail fast when one e2e test fails
The e2e tests are for a large part mutually exclusive, a failure of one
does not imply the others will fail. For this reason, it's useful to
have all e2e test jobs finish to potentially detect if something is
wrong related to the particular configuration of the matrix entry.