ci: add Action version in comments to Actions that use SHAs
This will allow Renovate to keep the SHAs pointing to tagged releases of
the corresponding Actions. It also makes the workflow a bit easier for
humans to understand (SHAs aren't human-readable, tag names (typically)
are).

The "renovate:" part of the comment was omitted, as RenovateBot removed
that when upgrading actions/checkout in [1]. This part of the comment
was also removed in `analysis.yml` where it was still present for some
Actions.

--
1. 910e41875645f843e4618172cb1da07addda282a