Add ShellCheck command and use it in CI
Add an npm command to run ShellCheck (configuration based on my
experience in other projects) and use it in the CI - i.e. run it in
GitHub Actions as well as the git hooks IF it's installed on the system.

Note that it was intentionally not incorporated into the `npm run lint`
command because we don't want to force contributors to install
ShellCheck on their system.