chore(deps): remove renovate
Remove the npm dependency "renovate" because:

1. it accounts for 1/3 of all npm dependencies.
2. it introduced (A)GPL licensed packages into the dependency tree.

While 2 is currently not a concern (because of how these packages are
used), it's not ideal to have these in the dependency tree.

Ultimately, the dependency adds relatively little value (the Renovate
configuration is not updated regularly), so I feel these reasons are
sufficient to remove the dependency.