ci: only audit production dependencies for releases
When auditing a release (a ref starting with "v"), don't bother auditing
devDependencies, as these generally don't affect already released
software. On non-release refs we're still interested in these because we
generally still want to avoid vulnerabilities in all dependencies on
development branches.

The `--production` flag is used even though it's deprecated because on
v2 "better-npm-audit" is used instead, which does not yet support the
"--omit dev" option.