chore: add `audit:prod` script
Add a new script similar to the audit script that is supposed to be a
version agnostic script for auditing production dependencies only.

As a reminder, the purpose of having an audit script is that the
implementation can differ from ref to ref (primarily relevant between
different major versions) while having a universal way of invoking
auditing in, e.g., CI (but also not having to remember the details when
switching between branches locally).