chore(deps): use Stryker jest-runner
Add the Stryker Mutator Jest runner as a dependency and update the
Stryker configuration to use it. Notable is the `testMatch` Jest option
seen in the Stryker configuration which ensures only the unit tests are
ran. Other configuration just follows the current recommendations from
the Jest runner docs.

For this to work, the Jest configuration had to be reverted to CommonJS
(otherwise importing that file from Stryker fails). And as a result of
that conversion the ESLint configuration had to be updated to turn off a
TypeScript specific rule for (cjs) configuration files.