chore: update incremental report location
Update the Stryker configuration to change the location of the
incremental report from the _reports/ dir to the .cache/ dir. This is
done because the incremental report isn't intended for human consumption
and it is used as a cache of sorts - improving the speed of subsequent
runs of Stryker.

The cache folder is called "stryker" rather than "mutation" because the
caching it does is specific to Stryker, a different mutation testing
framework would most likely use a different caching format.