chore: run and configure licensee
Add an npm script to run licensee:

    npm run check-licenses

And add a configuration file to configure licensee. For now, the
configuration just lists approved licenses (based on licenses of current
dependencies). This leaves a few "NOT APPROVED" licenses due to
dependency with undetected licenses.

The `"corrections"` option is disabled intentionally. This option allows
for letting licensee check packages against a (public) list of license
corrections for packages. While convenient, I prefer manual review for
now.