feat: add support for SVGO v3
Add support for SVGO v3 to the svgo module. The implementation
re-uses the wrapper for SVGO v2 as they're compatible. It's kept
separate to simplify changes when they become incompatible or support
for v2 is dropped.

The implementation passes the previously created integration tests and
is accompanied by new unit tests.