chore: setup hadolint-based Dockerfile linting
Add a script to run hadolint [1] on the devcontainer Dockerfile for this
project. Document this command in the Contributing Guidelines (note that
no new prerequisite is needed as hadolint is being invoked through
Docker). Add an initial configuration file for hadolint. Lint the file
continuously through git hooks and a step in the CI job "Check / Lint".

--
1. https://github.com/hadolint/hadolint