chore: enable `switch-exhaustiveness-check` rule
Enable the `switch-exhaustiveness-check` rule from
`(at)typescript-eslint`. This will ensure that all switch-statements in
the project are exhaustive, meaning every possible branch for the type
of the value being switched on is covered.

This will both avoid bugs (due to uncovered branches) as well as
possibly simplify code (removing unneeded branches).