test: update unit suite to no longer have implicit `any`s
Update all unit tests to avoid the usage of implicit `any`s. In some
cases this was as simple as defining the type. In many cases some type
inference (e.g. using `ReturnType<>`) was necessary, and so was casting
to `jest.MockedXxx<>` variants of types. In a couple of instances, the
required type is unnecessarily elaborate (when the type definition comes
from outside this project), in which case the values is cast for
simplicity.

Relatedly, some mocks had to be updated to not use implicit `any`s.

For some tests, a configuration object (or similar) type was necessary
in the test in order to guarantee test data has the correct type. This
required exposing the respective types from the source code as well as
the introduction of the `Mutable<>` type helper (similar to TypeScript
utility types [1]) to un-`readonly` those types to allow for mutations
during testing.

There's also some corrections:
- `__mocks__/(at)actions/core.ts`: correct mock name for getBooleanInput
- `src/types.d.ts`: correct InputterOptions.required type signature

--
1. https://www.typescriptlang.org/docs/handbook/utility-types.html