chore: create `npm run verify` command
Create a new command called "verify" that can be used to easily run all
the most important checks to verify changes are (probably) fine.
Recommend using this command in the Contributing Guidelines for the
contributor's convenience.

As this new command also audits the changes the text in the Contributing
Guidelines is updated to reflect this. This makes sense as all changes
should ideally be audited anyway.