chore: create asdf .tool-versions file
Create a `.tool-versions` file with less common system dependencies.
This file is first a configuration format for asdf [1], but is also
human-readable and thus can be used to know and track what versions of
dependencies _should_ be used.

Using asdf with this tool-versions file requires installing plugins for
act [2], actionlint [3], and ShellCheck [4].

--
1. https://asdf-vm.com/
2. https://github.com/grimoh/asdf-act
3. https://github.com/crazy-matt/asdf-actionlint
4. https://github.com/luizm/asdf-shellcheck