ci: use asdf Action for reproducibility w.r.t. tooling
Update the "Check / Lint" job to use asdf to install tooling (actionlint
and shellcheck are relevant in this case) to ensure the same versions of
these tools are used every CI run in order to improve reproducibility.

The harden-runner egress-policy has been set to "audit" in order to
re-evaluate allowed endpoints with the above change.