ci: create Action signature artifact for releases
Update the publish workflow to automatically create a signature of the
Action "executable" upon a release using cosign's [1] `sign-blob`
functionality.

To be able to do this, cosign is installed (pinned to a specific version
more on this in a bit). Prior to the creation of the GitHub release,
cosign is invoked with the `sign-blob` subcommand to sign the file
`lib/index.cjs` using "keyless signing" [2]. The output of this, the
signature and ephemeral key certificate, are attached to the release so
that users can verify the signature using the command:

    COSIGN_EXPERIMENTAL=1 \
    cosign verify-blob \
      --certificate index.cjs.cert \
      --signature index.cjs.sign \
      lib/index.cjs

This verification is included in the job to ensure it succeeds (and also
as a way of implicitly documenting how to perform the verification).

Note that "keyless signing" [2] is an experimental feature, hence the
`COSIGN_EXPERIMENTAL=1` environment variable and the pinning of cosign
to a specific version (to avoid running into breaking changes
unexpectedly).

--
1. https://github.com/sigstore/cosign
2. https://github.com/sigstore/cosign/blob/5282e31/KEYLESS.md