ci: split publish workflow into three jobs
Update the `publish.yml` workflow to consist of three distinct jobs. The
first one is just for checking if a release needs to be created for a
given commit on the `main` branch. The other two are for updating git
refs (tag and branch) creating a GitHub release respectively.

This change aims to simplify jobs, reduce the scope of available
permissions, aid in debugging, and reduce duplication (e.g. repeated if
statements).