ci: set up nightly tooling update job
Create a new job in the `nightly.yml` workflow that attempts to update
all tools listed in the `.tool-versions` file[^1] using `asdf`[1]. The
job is matrixed over the tools, will attempt to get the latest version
of the tool, install it locally, and create a Pull Request with the new
version (only created if there's actually a diff).

The "Get latest version" step is retried due to previous experiences
with this setup where that step could be a bit flaky.

--
^1: The list of tools is currently duplicated between the workflow and
the `.tool-versions` file. Resolving this duplication is an improvement
left for later.

--
1. https://asdf-vm.com/