ci: harden `release.yml` workflow
Prevent shell injection in this workflow by capturing the update type as
an environment variable and using the environment variable. This way, the
expansion of the input in the command can't result in shell injection.

Yes, this particular case is not very vulnerable due to 1) limited value
space, and 2) trusted triggers. However, it's hard to guarantee those
variables stay fixed and simple to apply the fix.

This problem was originally detected by Semgrep (semgrep.dev).