ci: update egress policy for "Check / CodeQL"
Update the "Check / CodeQL" job's egress policy to include two more
endpoints. This follows recent failed runs of the job where it was not
able to download the CodeQL CLI, due to the missing endpoint
`objects.githubusercontent.com`.

The other endpoint was added as a precaution, this is an official
endpoint used by GitHub Actions. It may or may not cause the job to
fail, but will in most likelihood show up as blocked at some point,
which isn't ideal.