ci: set up continuous SAST with DevSkim
Add a new CI job that runs the SAST tool DevSkim[1]. A tool by Microsoft
that:

> [...] is a security linter that highlights common security issues in
> source code.

The Action is based on a tool with the same name: DevSkim[2]. This is
also available as a CLI (not tried as of this commit). The tool supports
various languages, including for this project TypeScript.

This is mainly intended as a test for the tool.

--
1. https://github.com/microsoft/DevSkim-Action
2. https://github.com/Microsoft/DevSkim