ci: create job to track the size of `lib/index.cjs`
Create a new continuous integration job to track the size of the
`lib/index.cjs` file to start getting a handle on how this varies over
time.

The chosen Action was simply one of the first I found on the Actions
marketplace. The configuration is based on example workflows from this
Action's documentation.