ci: move Semgrep into separate workflow
Separate the continuous Semgrep job from the general checks workflow.
The motivation for this is that Semgrep requires access to a secret to
work, and this secret isn't available on Pull Requests from forks. This
would lead in a failed pipeline for external contributors, which is not
desired. Hence, just run it on `main` and address problems it finds
separately.