ci: set up a monthly CI job to update transitive dependencies
Create a new GitHub Actions workflow/job to update all transitive
dependencies on a weekly basis by running `npm update` (and `npm
dedupe`) and creating a Pull Request for the resulting changes.

This is experimental, intended to see if and how it works and if
there's any noticeable benefits.

It is set to run on the 23th day of the month which is based off of
the Renovate configuration to run on the first day of the month. This
gives 22 days to apply direct dependency updates and 5-8 days to apply
transitive dependency updates. Given the latter is typically easier, it
requires a smaller window to merge.