chore(deps): replace depcheck and unimported by knip
Add knip[1] as a devDependency replacing depcheck[2] and unimported[3].
knip provides the same functionality as those two tools combined, and a
bit more. The configuration for knip is largely based on that of
depcheck and unimported. Notable differences are:
- Ignore scripts as they're not of interest.
- Ignore binaries lists everything from the `.tool-versions` file (this
  wasn't covered by depcheck nor unimported).
- Plain "svgo" is ignored as a dependency, it wasn't detected by either
  dpecheck or unimported but is imported dynamically in
  `src/svgo/project.ts`. This is intentional as we want to be import a
  globally installed version of SVGO here.

The licensee configuration has been updated with 3 new exceptions due
to dependencies introduced in the knip dependency tree:
- deep-freeze: terms described as "public domain" [^1].
- filter-iterator: MIT license, see "License" section of the readme.
- identity-function: terms described as "public domain" [^1].

As of this commit the report by knip isn't quite clean yet, further
effort is required to get started actually using knip in practice.

--
1. https://github.com/webpro/knip
2. https://github.com/depcheck/depcheck
3. https://www.npmjs.com/package/unimported